requirements.txt
pyspark==3.5.0
apache-airflow==2.7.0
google-cloud-bigquery==3.7.0
google-cloud-storage==2.12.0
requests==2.31.0
pandas==2.1.0
apache-iceberg-spark3-runtime==1.3.1
apache-iceberg-gcp==1.3.1
python-dotenv==1.0.1

